Added datepicker component into gatheringinfo page. I moved the functions from ItineraryPage to the date picker component. I also rerouted the start button to the map page.
