+++ frontend/src/components/GatheringInfo/SearchBar.css
@@ -1,7 +1,7 @@
 
 .search-input {
-  width: 75%;
+  width: 35%;
   padding: 10px;
   border-radius: 105px;
-  border: 1px solid #ccc;
+  border: 2px solid #ccc;
 }
